controller.js is created and refactored
